requirements.txt
torch
torchvision
Pillow
gradio
requests
